% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idwas.R
\name{idwas}
\alias{idwas}
\title{Given an input of expression data and genomics data for a set of tumors, run IDWAS.}
\usage{
idwas(clinicalExpressionData, clinicalGenomicsData,
  clinicalCoVariatesMatrix = NULL, predictionLibrary = "pRRophetic",
  predictionFunction = "pRRopheticPredict", predictableDrugs = c("ABT.263",
  "ABT.888", "AG.014699", "AICAR", "ATRA", "Axitinib", "AZ628", "AZD.0530",
  "AZD.2281", "AZD6244", "AZD6482", "AZD7762", "BAY.61.3606", "BIBW2992",
  "Bicalutamide", "BI.D1870", "Bleomycin", "BMS.536924", "BMS.754807",
  "Bortezomib", "Bosutinib", "BX.795", "Camptothecin", "CEP.701", "CGP.082996",
  "CHIR.99021", "CI.1040", "Cisplatin", "Cytarabine", "Dasatinib", "DMOG",
  "Docetaxel", "Elesclomol", "Erlotinib", "FH535", "FTI.277", "Gefitinib",
  "Gemcitabine", "IPA.3", "Lapatinib", "Methotrexate",      "MG.132",
  "Midostaurin", "Mitomycin.C", "Nilotinib", "Nutlin.3a", "NVP.BEZ235",
  "NVP.TAE684", "Obatoclax.Mesylate", "PAC.1", "PD.0325901", "PD.0332991",
  "PD.173074", "PLX4720", "RDEA119", "SB590885", "Sunitinib", "Temsirolimus",
  "Thapsigargin", "Tipifarnib", "TW.37", "Vinblastine", "Vorinostat", "VX.702",
  "WH.4.023", "WO2009093972", "WZ.1.84", "X17.AAG", "X681640", "XMD8.85",
  "ZM.447439"), ...)
}
\arguments{
\item{clinicalExpressionData}{Numeric matrix of tumor gene expresison data. Colum names must be patient IDs, row names must be HUGO gene symbols.}

\item{clinicalGenomicsData}{Numeric matrix representing another source of data that will be comapred to imputed drug response. E.g. genomics data, whereby a mutated gene as been coded as a 1 and wt as 0. Rownames must identify the abberation (e.g. gene symbols) and column names must represent patients. Column names must be the same as the gene expression data.}

\item{clinicalCoVariatesMatrix}{A matrix of co-variates to included in the association analysis, e.g. cancer type or GLDS.}

\item{predictionLibrary}{What should be used to impute drug response. Currently, pRRophetic is supported, but in theory any prediction method could be implemented.}

\item{predictionFunction}{What function within the predictionLibrary should be used, currently pRRopheticPredict is supported.}

\item{predictableDrugs}{A vector of drugs for which we want to predict drug response. By default it is a list of drug with Spearman correlation of greater than 0.3 in cross-validation on the GDSC 2014 dataset.}
}
\value{
a list of matrices of p-values and effect sizes, for each drug against each aberration in clinicalGenomicsData.
}
\description{
This function takes as its input gene expression and genomics data from a set of cancer
patients' tumors. The purpose of this function is to impute response in these tumors
using the pRRophetic prediction library, then to identify variables in the genomics data
supplied that are correlated with predicted drug response.
}
\keyword{idwas}
\keyword{perform}
